change from PATTERNS of regex to TESTERS of functions
doing regex on the uri isn't flexible enough. one might also
want to get (or head) the uri and look at the content type
and from that determine which thing to use.

This came up while trying to work out how to do an atom feed
as a remote bag.